core: Introduce signal-safe backtrace facility
backwalk uses dladdr to resolve symbol addresses, which is not
signal-safe. This patch introduces two APIs that decouple the backtrace
collection from the processing phase. The collection API is
signal-safe and can be called from a signal handler, while the
processing API can be called later on a possibly separate thread. See
signal_test.c for example usage.